refactor: More cleanup - organize utilities
