+++ GPSTest/src/main/java/com/android/gpstest/util/UIUtils.kt
@@ -1103,4 +1103,31 @@ internal object UIUtils {
             PreferenceUtils.saveInt(WHATS_NEW_VER, appInfo.versionCode)
         }
     }
+
+    /**
+     * Returns the `location` object as a human readable string for use in a notification title
+     */
+    fun Location?.toNotificationTitle(): String {
+        return if (this != null) {
+            FormatUtils.formatLatOrLon(latitude, CoordinateType.LATITUDE) +
+                    "," +
+                    FormatUtils.formatLatOrLon(longitude, CoordinateType.LONGITUDE)
+        } else {
+            "Unknown location"
+        }
+    }
+
+    /**
+     * Returns the `location` object as a human readable string for use in a notification summary
+     */
+    fun Location?.toNotificationSummary(): String {
+        return if (this != null) {
+            val lat = FormatUtils.formatLatOrLon(latitude, CoordinateType.LATITUDE)
+            val lon = FormatUtils.formatLatOrLon(longitude, CoordinateType.LONGITUDE)
+            val alt = FormatUtils.formatAltitude(this)
+            "$lat $lon $alt"
+        } else {
+            "Unknown location"
+        }
+    }
 }
